fix(architecture): move scroll-progress-bar out of components/ui to comply with design system constraints
The hushh project strictly limits components/ui to approved design system contract files (like shadcn base components). This moves the custom scroll-progress-bar feature to components/app-ui to resolve the verify:design-system CI failure.
