+++ hushh-webapp/components/app-ui/scroll-progress-bar.tsx
